refactor(api): move message API contracts to API crate
Move Message, MessageRole, ContentPart, CreateMessageRequest and
related types from everruns-contracts to everruns-api/src/messages.rs.
This keeps API-specific contracts in the API crate while runtime
types remain in everruns-core.
