+++ crates/everruns-contracts/src/lib.rs
@@ -1,13 +1,14 @@
 // Public contracts for Everruns API
 // This crate defines DTOs, AG-UI event types, and JSON schemas
 // M2: Agent/Session/Messages model with Events as SSE notification channel
+//
+// Note: Message API contracts have been moved to everruns-api crate
 
 pub mod agent;
 pub mod capability;
 pub mod common;
 pub mod events;
 pub mod llm;
-pub mod message;
 pub mod session;
 pub mod tools;
 
@@ -16,6 +17,5 @@ pub use capability::*;
 pub use common::*;
 pub use events::*;
 pub use llm::*;
-pub use message::*;
 pub use session::*;
 pub use tools::*;
